Simplify _compute_local_shape_and_global_offset and make it SPMD.
There is only one substantive change: the branch on
`global_offset[shard_dim] <= local_offset[shard_dim]`
is removed because it is unnecessary: you can always treat the
first shard uniformly with the rest of the shards, because your
global offset is guaranteed to be zero in this case anyway.

I also switch the shard_size case to sym_ite, to make it possible
for LocalTensor to deal with the MPMD-ness here, but it's equivalent
to the old if-then-else.

I tried to rewrite the comments to be more clear what is going on
algorithmically here.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: 5a53fea5695f87e3e0908365b89696e19e9a0d72
Pull Request resolved: https://github.com/pytorch/pytorch/pull/163344