[AOTI] Pass comments from metadata to the autotune block
Summary: When generating Triton kernels in the compile-time autotune blocks, it will be useful to generate source information as code comments. Previously we ignore these comments for autotune code blocks because the generated main output code will contain the same information, but it won't work if the generated autotune code crashes.

ghstack-source-id: 9d5fc0fe93ce890eebf16eabcba0f05415b11a16
Pull Request resolved: https://github.com/pytorch/pytorch/pull/163600